Separate index.js and bundle instance for documentation purposes/clarity
